settings: Don't default repo visibility to `public` (#6)
This setting automatically sets new repositories in the organization
to `public`. It does not matter if the repo was explicitly set to
`internal` or `private.` The bot changes the repo to `public`. It
can easily cause proprietary data to be exposed if the creator of
the repo doesn't catch it in time before the first push.

Signed-off-by: Stephen Augustus <foo@auggie.dev>